+++ tests/FormBuilderTestBase.php
@@ -1,47 +1,47 @@
 <?php
 use A2design\Form\FormBuilder;
 
-class FormBuilderTest extends Orchestra\Testbench\TestCase
+class FormBuilderTestBase extends Orchestra\Testbench\TestCase
 {
     /**
      * @var \A2design\Form\FormBuilder
      */
-    private $formBuilder;
+    protected $formBuilder;
 
     /**
      * Mock of \Illuminate\View\Factory
      *
      * @var \Mockery\MockInterface
      */
-    private $viewFactory;
+    protected $viewFactory;
 
     /**
      * Mock of \Illuminate\Session\Store
      *
      * @var \Mockery\MockInterface
      */
-    private $session;
+    protected $session;
 
     /**
      * Mock of \Illuminate\Config\Repository
      *
      * @var \Mockery\MockInterface
      */
-    private $config;
+    protected $config;
 
     /**
      * Mock of \Illuminate\Routing\RouteCollection
      *
      * @var \Mockery\MockInterface
      */
-    private $routes;
+    protected $routes;
 
     /**
      * Mock of \Illuminate\Http\Request
      *
      * @var \Mockery\MockInterface
      */
-    private $request;
+    protected $request;
 
     /**
      * Specify the package service provider
@@ -109,92 +109,6 @@ class FormBuilderTest extends Orchestra\Testbench\TestCase
         Mockery::close();
     }
 
-    public function testOpening()
-    {
-        $defaultForm = $this->formBuilder->create();
-
-        $this->assertContains('<form', $defaultForm, 'Form is not opened');
-    }
-
-    public function testTokenInput()
-    {
-        $defaultForm = $this->formBuilder->create();
-        $message = 'Some issue with token hidden input';
-
-        $this->assertContains('type="hidden"', $defaultForm, $message);
-        $this->assertContains('name="_token"', $defaultForm, $message);
-        //we cant check the csrf_token value because in test case is null
-        $this->assertContains('value="', $defaultForm, $message);
-
-        $getForm = $this->formBuilder->create('', null, ['method' => 'get']);
-        $this->assertNotContains('type="hidden"', $getForm, $message);
-        $this->assertNotContains('name="_token"', $getForm, $message);
-
-        $getForm = $this->formBuilder->create('', null, ['method' => 'GET']);
-        $this->assertNotContains('type="hidden"', $getForm, $message);
-        $this->assertNotContains('name="_token"', $getForm, $message);
-    }
-
-    public function testMethodInput()
-    {
-        //
-    }
-
-    public function testMethodDetection()
-    {
-        //
-    }
-
-    public function testAction()
-    {
-        //
-    }
-
-    public function testId()
-    {
-        //
-    }
-
-    public function testClass()
-    {
-        //
-    }
-
-    public function testAbsoluteParameter()
-    {
-        //
-    }
-
-    public function testUrlParameter()
-    {
-        //
-    }
-
-    public function testFormDirectionParameter()
-    {
-        //
-    }
-
-    public function testAattrsParameter()
-    {
-        //
-    }
-
-    public function testEnctypeParameter()
-    {
-        //
-    }
-
-    public function testHasFilesParameter()
-    {
-        //
-    }
-
-    public function testParameterInheriting()
-    {
-        //
-    }
-    
     /**
      * Set initial config values for the config mock
      *
@@ -202,7 +116,7 @@ class FormBuilderTest extends Orchestra\Testbench\TestCase
      *
      * @return \Mockery\MockInterface
      */
-    private function stubConfig($configMock)
+    protected function stubConfig($configMock)
     {
         /** @var \Illuminate\Config\Repository $configs */
         $configs = config();
